requirements.txt
yt-dlp==2025.1.26
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
tqdm==4.66.4
pytest==8.3.3

brotlicffi>=1.1.0
pycryptodomex>=3.20
mutagen>=1.47
